+++ src/main/java/net/oneandone/pommes/cli/Status.java
@@ -17,16 +17,23 @@ package net.oneandone.pommes.cli;
 
 import net.oneandone.inline.ArgumentException;
 import net.oneandone.pommes.database.Database;
+import net.oneandone.pommes.database.Field;
 import net.oneandone.pommes.database.Project;
+import net.oneandone.pommes.descriptor.Descriptor;
+import net.oneandone.pommes.repository.NodeRepository;
 import net.oneandone.pommes.scm.Scm;
 import net.oneandone.sushi.fs.DirectoryNotFoundException;
 import net.oneandone.sushi.fs.ListException;
 import net.oneandone.sushi.fs.file.FileNode;
 import net.oneandone.sushi.fs.filter.Filter;
+import net.oneandone.sushi.util.Separator;
+import net.oneandone.sushi.util.Strings;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -41,37 +48,61 @@ public class Status extends Base {
     @Override
     public void run(Database database) throws Exception {
         Map<FileNode, Scm> checkouts;
+        Map<String, Step> steps;
+        int id;
+        String idStr;
         FileNode found;
-        FileNode expected;
-        Project foundProject;
-        Scm scm;
+        Step step;
 
         checkouts = Scm.scanCheckouts(directory, environment.excludes());
         if (checkouts.isEmpty()) {
             throw new ArgumentException("no checkouts in " + directory);
         }
+        steps = new LinkedHashMap<>();
+        id = 0;
         for (Map.Entry<FileNode, Scm> entry : checkouts.entrySet()) {
             found = entry.getKey();
-            scm = entry.getValue();
-            foundProject = database.projectByScm(scm.getUrl(found));
-            if (foundProject == null) {
-                console.info.println("? " + found);
+            step = Step.create(environment, database, found, entry.getValue());
+            if (step.isNoop()) {
+                idStr = " ";
             } else {
-                try {
-                    expected = environment.home.root().directory(foundProject);
-                } catch (IOException e) {
-                    console.info.println("! " + found + " " + e.getMessage());
-                    continue;
-                }
-                if (found.equals(expected)) {
-                    console.info.println((scm.isCommitted(found) ? ' ' : 'M') + " " + found);
-                } else {
-                    console.info.println("C " + found);
-                }
+                idStr = Integer.toString(++id);
+                steps.put(idStr, step);
             }
+            console.info.println(step.toString(idStr));
         }
         for (FileNode u : unknown(directory, checkouts.keySet(), environment.excludes())) {
             console.info.println("? " + u + " (normal directory)");
+        }
+
+        if (!steps.isEmpty()) {
+            while (true) {
+                String input;
+                input = console.readline("What do you want to fix, ctrl-c to abort (<numbers>/all)? ");
+                if ("all".equals(input)) {
+                    for (var entry : steps.entrySet()) {
+                        entry.getValue().apply(database);
+                        console.info.println("fixed " + entry.getKey());
+                    }
+                    steps.clear();
+                } else {
+                    for (String item : Separator.SPACE.split(input)) {
+                        step = steps.remove(item);
+                        if (step == null) {
+                            console.info.println("unknown input: " + item);
+                            break;
+                        }
+                        step.apply(database);
+                        console.info.println("fixed " + item);
+                    }
+                }
+                if (steps.isEmpty()) {
+                    break;
+                }
+                for (var entry : steps.entrySet()) {
+                    console.info.println(entry.getValue().toString(entry.getKey()));
+                }
+            }
         }
     }
 
@@ -118,6 +149,122 @@ public class Status extends Base {
                 result.add(parent);
                 candicates(root, parent, result);
             }
+        }
+    }
+
+    public static class Step {
+        public static Step create(Environment environment, Database database, FileNode found, Scm scm) throws IOException {
+            Project foundProject;
+            String scmUrl;
+            Project newPom;
+            Descriptor probed;
+            FileNode expected;
+            Relocation relocation;
+
+            scmUrl = scm.getUrl(found);
+            foundProject = database.projectByScm(scmUrl);
+            if (foundProject == null) {
+                probed = NodeRepository.probe(environment, found);
+                if (probed == null) {
+                    throw new IllegalStateException();
+                }
+                newPom = probed.load(environment, "local");
+                expected = environment.home.root().directory(newPom);
+                relocation = found.equals(expected)? null : new Relocation(found, expected);
+                return new Step("?", found.toString(), scmUrl, newPom, relocation);
+            } else {
+                try {
+                    expected = environment.home.root().directory(foundProject);
+                } catch (IOException e) {
+                    return new Step("!", found.toString() + " " + e.getMessage(), scmUrl, null, null);
+                }
+                if (found.equals(expected)) {
+                    return new Step(scm.isCommitted(found) ? " " : "M", found.toString(), scmUrl, null, null);
+                }
+                return new Step("C", found.toString(), scmUrl, null, new Relocation(found, expected));
+            }
+        }
+
+        private final String marker;
+        private final String message;
+        private final String scmUrl;
+        private final Project newPom;
+        private final Relocation relocation;
+
+        public Step(String marker, String message, String scmUrl, Project newPom, Relocation relocation) {
+            this.marker = marker;
+            this.message = message;
+            this.scmUrl = scmUrl;
+            this.newPom = newPom;
+            this.relocation = relocation;
+        }
+
+        public boolean isNoop() {
+            return newPom == null && relocation == null;
+        }
+
+        public void apply(Database database) throws IOException {
+            if (newPom != null) {
+                database.index(Collections.singleton(Field.document(newPom)).iterator());
+            }
+            if (relocation != null) {
+                relocation.apply();
+            }
+        }
+
+        public String toString(String id) {
+            String head = "[" + id + "]" + " " + marker + " ";
+            List<String> lines;
+            StringBuilder result;
+
+            lines = new ArrayList<>();
+            lines.add(message);
+            if (newPom != null) {
+                lines.add("add " + scmUrl);
+            }
+            if (relocation != null) {
+                lines.add(relocation.toString());
+            }
+            if (lines.isEmpty()) {
+                throw new IllegalStateException();
+            }
+            result = new StringBuilder();
+            boolean first = true;
+            for (String line : lines) {
+                if (first) {
+                    first = false;
+                    result.append(head);
+                } else {
+                    result.append('\n');
+                    result.append(Strings.times(' ', head.length()));
+                }
+                result.append(line);
+            }
+            return result.toString();
+        }
+    }
+
+    public static class Relocation {
+        private final FileNode found;
+        private final FileNode expected;
+
+        public Relocation(FileNode found, FileNode expected) {
+            this.found = found;
+            this.expected = expected;
+        }
+
+        public void apply() throws IOException {
+            expected.getParent().mkdirsOpt();
+            found.move(expected);
+        }
+
+        public String toString() {
+            FileNode parent;
+            String mkdir;
+
+            parent = expected.getParent();
+            mkdir = parent.exists() ? "" : "mkdir -p " + parent + "; ";
+            return "relocate: " + mkdir + "mv " + found + " " + expected;
         }
     }
 }