+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * added `maintenance` command to automatically add and relocate unknown projects or relocates existing project; 
   status no longer shows a fix for that 
+* status provements
+  * sort entries
 * added "raw projects" that represent projects without meta information
 * changed checkout paths: they are now based on the url, not gav. That feels more natural, and it works
   for all projects, even without pom