+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   for all projects, even without pom
 * support shorter scp-like git ssh urls (used by gitea)
 * improved error message for git.getUrl
+* rename `master` branch to `main`
 * update sushi 3.2.0 to 3.3.0
 * update gson 2.8.0 to 2.8.5
 * update lazy-foss-parent 1.5.0 to 1.5.1