+++ CHANGELOG.md
@@ -2,11 +2,10 @@
 
 ### 3.3.0 (pending)
 
-* added `maintenance` command to automatically add and relocate unknown projects or relocates existing project; 
-  status no longer shows a fix for that 
 * status improvements
   * sort entries
   * don't print the project, just the checkout
+  * interactively offer to add/relocate where possible
 * added "raw projects" that represent projects without meta information
 * changed checkout paths: they are now based on the url, not gav. That feels more natural, and it works
   for all projects, even without pom